requirements.txt
prometheus_client>=0.17.0
psycopg2-binary>=2.9.9
python-dateutil>=2.8.2
